pyproject.toml
[project]
name = "catclaw"
version = "0.1.0"
description = "猫爪"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.12.0",
    "open-skills>=0.2.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
